prod-devel-rcar.yaml: Add meta-xt-vhost layer
To be able to use vhost_vsock and other vhost functionality under the
Xen hypervisor, we need to apply a set of patches on the DomD kernel.
The target patches are accumulated inside the 'meta-xt-vhost' layer.

This patch is adding usage of the 'meta-xt-vhost' layer to the driver
domain in case of virtio builds.

Please note, this layer is private (at least for now). This fact will
break the virtio build for anyone who has no access to that repo,
the suggestion is to remove "../meta-xt-vhost" inclusion from domd's
layers in prod-devel-rcar-virtio.yaml

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>
Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>