domd, domu: Add installation of the iperf tool
We need iperf tool to be able to perform network throughput
measurements. E.g. we want to test whether the network will be faster
if the network device, shared by QEMU from DomD to DomU works through
the 'vhost'.

We've decided to permanently install this tool in DomD and DomU, as
such kind of testing might be needed from time to time in different
contexts. The benefits of having this tool permanently available are
higher than the additional space, that it consumes.

Signed-off-by: Vladyslav Goncharuk <vladyslav_goncharuk@epam.com>